# Minor all-gather optimization