--- a/models/demos/llama3_subdevices/tests/tg_perf_unit_tests/test_ccl_async_perf_TG_llama.py
+++ b/models/demos/llama3_subdevices/tests/tg_perf_unit_tests/test_ccl_async_perf_TG_llama.py
@@ -14,9 +14,9 @@ THRESHOLD = 0.4
 @pytest.mark.parametrize(
     "ag_type, warmup_iters, perf_target_us",
     [
-        ("sdpa", 15, 9.49),
-        ("binary_mult", 15, 10.54),
-        ("layernorm", 15, 6.47),
+        ("sdpa", 15, 9.01),
+        ("binary_mult", 15, 10.10),
+        ("layernorm", 15, 6.22),
     ],
 )
 @pytest.mark.models_device_performance_bare_metal