Use a more advanced two step fabric connection close in all-gather for some additional minor time savings (couple hundred ns per op).

This was missed in this earlier PR: https://github.com/tenstorrent/tt-metal/pull/19715

Results:
| Op | Before (us) | After (us) | Delta (us) |
|-|-|-|-|
| All-gather binary mult | 10.50  | 10.10 | 0.4 |
| All-gather SDPA | 9.49 | 9.00 | 0.49|
| All-gather layernorm | 6.47 | 6.32 | 0.15 |


### Checklist
- [ ] [All post commit](https://github.com/tenstorrent/tt-metal/actions/workflows/all-post-commit-workflows.yaml) CI passes
- [ ] [Blackhole Post commit](https://github.com/tenstorrent/tt-metal/actions/workflows/blackhole-post-commit.yaml) CI passes (if applicable)
- [ ] [Model regression](https://github.com/tenstorrent/tt-metal/actions/workflows/perf-models.yaml) CI passes (if applicable)
- [ ] [Device performance regression](https://github.com/tenstorrent/tt-metal/actions/workflows/perf-device-models.yaml) CI passes (if applicable)
- [ ] **(For models and ops writers)** Full [new models tests](https://github.com/tenstorrent/tt-metal/actions/workflows/full-new-models-suite.yaml) CI passes (if applicable)
- [ ] New/Existing tests provide coverage for changes